Change forkserver and cmplog to handle exec_tmout = 0

+++ src/afl-fuzz-cmplog.c
@@ -182,19 +182,25 @@ void init_cmplog_forkserver(afl_state_t *afl) {
 
   /* Wait for the fork server to come up, but don't wait too long. */
 
-  fd_set readfds;
+  if (afl->fsrv.exec_tmout) {
 
-  FD_ZERO(&readfds);
-  FD_SET(afl->cmplog_fsrv_st_fd, &readfds);
-  timeout.tv_sec = ((afl->fsrv.exec_tmout * FORK_WAIT_MULT) / 1000);
-  timeout.tv_usec = ((afl->fsrv.exec_tmout * FORK_WAIT_MULT) % 1000) * 1000;
+    fd_set readfds;
+    FD_ZERO(&readfds);
+    FD_SET(afl->cmplog_fsrv_st_fd, &readfds);
+    timeout.tv_sec = ((afl->fsrv.exec_tmout * FORK_WAIT_MULT) / 1000);
+    timeout.tv_usec = ((afl->fsrv.exec_tmout * FORK_WAIT_MULT) % 1000) * 1000;
 
-  int sret = select(afl->cmplog_fsrv_st_fd + 1, &readfds, NULL, NULL, &timeout);
+    int sret = select(afl->cmplog_fsrv_st_fd + 1, &readfds, NULL, NULL, &timeout);
 
-  if (sret == 0) {
+    if (sret == 0) {
 
-    kill(afl->cmplog_fsrv_pid, SIGKILL);
+      kill(afl->cmplog_fsrv_pid, SIGKILL);
 
+    } else {
+
+      rlen = read(afl->cmplog_fsrv_st_fd, &status, 4);
+
+    }
   } else {
 
     rlen = read(afl->cmplog_fsrv_st_fd, &status, 4);
